requirements.txt
pymupdf
pytesseract
Pillow
scikit-learn
numpy
scipy
pdf2image